Error handlers

+++ src/controllers/bookings/deleteBooking.js
@@ -1,15 +1,21 @@
 import { PrismaClient } from "@prisma/client";
-import notFoundError from "../../errors/notFoundError.js";
+import NotFoundError from "../../errors/NotFoundError.js";
+import BadRequestError from "../../errors/BadRequestError.js";
 
 const deleteBooking = async (id) => {
   const prisma = new PrismaClient();
+
+  if (!id) {
+    throw new BadRequestError("id is required");
+  }
+
   const booking = await prisma.booking.deleteMany({
     where: {
       id: id,
     },
   });
   if (!booking || booking.count === 0) {
-    throw new notFoundError("Booking", id);
+    throw new NotFoundError("Booking", id);
   }
   return id;
 };
